live share 08/09

+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/AlquilerRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Alquiler;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AlquilerRepository extends JpaRepository<Alquiler, Long> {
+public interface AlquilerRepository extends JpaRepository<Alquiler, Integer> {
 }
