+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/LocalidadRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Localidad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocalidadRepository extends JpaRepository<Localidad, Long> {
+public interface LocalidadRepository extends JpaRepository<Localidad, Integer> {
 }
