+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/PagoRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Pago;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PagoRepository extends JpaRepository<Pago, Long> {
+public interface PagoRepository extends JpaRepository<Pago, Integer> {
 }
