+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/ProvinciaRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.Provincia;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProvinciaRepository extends JpaRepository<Provincia, Long> {
+public interface ProvinciaRepository extends JpaRepository<Provincia, Integer> {
 }
