+++ src/main/java/com/seminario/gestorInmobiliario/Repositorios/VisitaPropiedadRepository.java
@@ -3,5 +3,5 @@ package com.seminario.gestorInmobiliario.Repositorios;
 import com.seminario.gestorInmobiliario.Entidades.VisitaPropiedad;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VisitaPropiedadRepository extends JpaRepository<VisitaPropiedad, Long> {
+public interface VisitaPropiedadRepository extends JpaRepository<VisitaPropiedad, Integer> {
 }
